requirements.txt
jira
python-dotenv
urllib3